Update messages and message_types migration files to improve data structure. Change type_id from unsignedBigInteger to unsignedTinyInteger for better data efficiency. Modify message_type table to use tinyIncrements for id, enhancing primary key management.

+++ database/migrations/2024_01_01_000001_create_messages_table.php
@@ -8,10 +8,11 @@ return new class extends Migration
 {
     public function up()
     {
-        Schema::create('messages', function (Blueprint $table) {
+        Schema::create('messages', function (Blueprint $table)
+        {
             $table->id();
             $table->string('name');
-            $table->unsignedBigInteger('type_id');
+            $table->unsignedTinyInteger('type_id');
             $table->unsignedBigInteger('category_id')->nullable();
             $table->unsignedBigInteger('contact_status_id')->nullable();
             $table->unsignedBigInteger('template_id')->nullable();
